backend/requirements.txt
fastapi
uvicorn[standard]
pydantic-settings
python-jose[cryptography]
passlib[bcrypt]
openai
python-multipart
websockets==12.0
Authlib